新增企业、企业用户关系

+++ user-center/user-api/src/main/java/com/lcdt/userinfo/service/UserService.java
@@ -16,7 +16,7 @@ public interface UserService {
 
 	boolean checkUserLogin();
 
-	FrontUserInfo queryByUserId(Integer userId) throws UserNotExistException;
+	FrontUserInfo queryByUserId(Long userId) throws UserNotExistException;
 
 	FrontUserInfo queryByPhone(String phone) throws UserNotExistException;
 }
